update composer role
add module to install global composer packages

+++ plugins/modules/composer_installer.py
@@ -7,6 +7,7 @@ from __future__ import absolute_import, print_function
 import urllib3
 import requests
 import os
+import re
 
 from pathlib import Path
 
@@ -17,7 +18,7 @@ from ansible_collections.bodsch.core.plugins.module_utils.directory import creat
 from ansible_collections.bodsch.core.plugins.module_utils.file import create_link
 
 
-class PHPComposerInstaller(object):
+class ComposerInstaller(object):
     """
     """
     module = None
@@ -34,10 +35,11 @@ class PHPComposerInstaller(object):
         self.version_branch = module.params.get("version_branch")
         self.update = module.params.get("update")
         self.path = module.params.get("path")
+        self.force = module.params.get("force", False)
 
         self.cache_directory = f"{Path.home()}/.ansible/composer"
         self.composer_installer = os.path.join(self.cache_directory, "composer-installer.php")
-        self.composer_signature = os.path.join(self.cache_directory, "composer.sha384")
+        self.composer_signature = os.path.join(self.cache_directory, "composer-installer.sha384")
 
         self.php_bin = module.get_bin_path("php", True)
         self.composer_bin = module.get_bin_path("composer", False)
@@ -50,44 +52,36 @@ class PHPComposerInstaller(object):
     def run(self):
         result = dict(
             failed=True,
-            available_php_version="none"
+            msg="De wereld is om zeep."
         )
 
         create_directory(self.cache_directory)
 
-        self.module.log(msg=f" - {self.composer_installer}")
-        self.module.log(msg=f" - {self.composer_signature}")
-        self.module.log(msg=f" - {self.composer_bin}")
-
         self.checksum = Checksum(self.module)
 
-        # if not os.path.exists(self.composer_installer) and os.path.exists(self.composer_signature):
-        #     os.remove(self.composer_signature)
-        # elif not os.path.exists(self.composer_signature) and os.path.exists(self.composer_installer):
-        #     os.remove(self.composer_installer)
+        if bool(self.force):
+            # self.module.log(msg=f"force mode: {self.force} {type(self.force)}")
+            self.remove_file(self.composer_signature)
+            self.remove_file(self.composer_installer)
+            self.remove_file(self.composer_bin)
+            self.remove_file(self.composer_bin_checksum)
 
+        """
+            check cached signature
+        """
         if os.path.exists(self.composer_signature):
             out_of_cache_sign = cache_valid(self.module, cache_file_name=self.composer_signature, cache_minutes=self.cache_minutes)
-            # out_of_cache_inst = cache_valid(self.module, cache_file_name=self.composer_installer, cache_minutes=self.cache_minutes)
 
             if not out_of_cache_sign:
-                (changed, checksum_from_file, old_checksum) = self.validate_checksum()
-
-                if changed:
-                    self.__remove_file(self.composer_signature)
-                    self.__remove_file(self.composer_installer)
+                self.validate_checksum(remove_changed=True)
 
         if not os.path.exists(self.composer_signature) or not os.path.exists(self.composer_installer):
             """
             """
             self.download_installer_files()
-            (changed, checksum_from_file, old_checksum) = self.validate_checksum()
+            self.validate_checksum(remove_changed=True)
 
-            if changed:
-                self.__remove_file(self.composer_signature)
-                self.__remove_file(self.composer_installer)
-
-        else:
+        if os.path.exists(self.composer_installer):
             result = self.run_installer()
 
         return result
@@ -95,25 +89,21 @@ class PHPComposerInstaller(object):
     def download_installer_files(self):
         """
         """
-        self.module.log(msg=f"download_installer_files()")
-
-        status_code, output = self.__call_url(url=self.installer)
+        status_code, output = self.call_url(url=self.installer)
 
         if status_code == 200:
             with open(self.composer_installer, "w") as f:
                 f.write(output)
 
-        status_code, output = self.__call_url(url=self.signature)
+        status_code, output = self.call_url(url=self.signature)
 
         if status_code == 200:
             with open(self.composer_signature, "w") as f:
                 f.write(output)
 
-    def validate_checksum(self):
+    def validate_checksum(self, remove_changed=False):
         """
         """
-        self.module.log(msg=f"validate_checksum()")
-
         old_checksum = ""
 
         checksum_from_file = self.checksum.checksum_from_file(path=self.composer_installer, algorithm="sha384")
@@ -122,35 +112,59 @@ class PHPComposerInstaller(object):
             with open(self.composer_signature, "r") as f:
                 old_checksum = f.readlines()[0].strip()
 
-        self.module.log(msg=f" xxx checksum    : {checksum_from_file}")
-        self.module.log(msg=f" old checksum    : {old_checksum}")
-
         changed = not (old_checksum == checksum_from_file)
+
+        if remove_changed and changed:
+            self.remove_file(self.composer_signature)
+            self.remove_file(self.composer_installer)
 
         return (changed, checksum_from_file, old_checksum)
 
-    def __remove_file(self, filename):
-        self.module.log(msg=f"__remove_file({filename})")
+    def composer_version(self):
+        """
+        """
+        args = []
 
-        if os.path.exists(filename):
+        composer_version = ""
+        composer_date = ""
+
+        if self.composer_bin and os.path.exists(self.composer_bin):
+            args.append(self.php_bin)
+            args.append(self.composer_bin)
+            args.append("--version")
+
+            rc, out, err = self._exec(args)
+
+            if rc == 0:
+                pattern = re.compile(r'^Composer version (?P<version>[0-9\.]+) (?P<date>.+)$')
+
+                result = re.search(pattern, out)
+                composer_version = result.group('version')
+                composer_date = result.group('date')
+
+        return (composer_version, composer_date)
+
+    def remove_file(self, filename):
+        self.module.log(msg=f"remove_file({filename})")
+
+        if filename and os.path.exists(filename):
             os.remove(filename)
 
     def run_installer(self):
         """
         """
-        self.module.log(msg=f"run_installer()")
-
         args = []
 
         if self.composer_bin and os.path.exists(self.composer_bin):
             (changed, checksum_from_file, old_checksum) = self.checksum.validate_from_file(checksum_file=self.composer_bin_checksum, data_file=self.composer_bin)
 
+            composer_version, composer_date = self.composer_version()
+
             if not changed:
                 return dict(
-                    rc=0,
                     failed=False,
                     changed=False,
-                    version=self.version
+                    version=composer_version
                 )
         else:
             args.append(self.php_bin)
@@ -166,12 +180,9 @@ class PHPComposerInstaller(object):
             args.append("--install-dir")
             args.append(self.path)
 
-            self.module.log(msg=f" - args {args}")
-
             rc, out, err = self._exec(args)
 
             if rc == 0:
-
                 composer_bin_phar = self.module.get_bin_path("composer.phar", False)
                 composer_bin_dest = os.path.join(self.path, "composer")
 
@@ -181,30 +192,23 @@ class PHPComposerInstaller(object):
         self.composer_bin = self.module.get_bin_path("composer", False)
 
         if os.path.exists(self.composer_bin) and not os.path.exists(self.composer_bin_checksum):
-
             checksum = self.checksum.checksum_from_file(self.composer_bin)
             self.checksum.write_checksum(self.composer_bin_checksum, checksum)
 
+            composer_version, composer_date = self.composer_version()
+
         return dict(
-            rc=0,
             failed=False,
             changed=True,
-            version=self.version
+            version=composer_version
         )
 
-        return (rc, out, err)
-
-    def __call_url(self, url=None, method='GET', data=None):
+    def call_url(self, url=None, method='GET', data=None):
         """
         """
         response = None
 
         headers = {}
-        #    "Accept": "application/json",
-        #    "Content-Type": "application/json;charset=utf-8"
-        # }
-
-        # github_url = f"{self.github_url}/{self.version}/{self.checksum_file}"
 
         try:
             # authentication = (self.github_username, self.github_password)
@@ -222,36 +226,22 @@ class PHPComposerInstaller(object):
 
             response.raise_for_status()
 
-            # self.module.log(msg=f" text    : {response.text} / {type(response.text)}")
-            # self.module.log(msg=f" json    : {response.json()} / {type(response.json())}")
-            # self.module.log(msg=f" headers : {response.headers}")
-            # self.module.log(msg=f" code    : {response.status_code}")
-            # self.module.log(msg="------------------------------------------------------------------")
-
             return response.status_code, response.text
 
         except requests.exceptions.HTTPError as e:
             self.module.log(msg=f"ERROR   : {e}")
             status_code = e.response.status_code
             status_message = e.response.text
-            # self.module.log(msg=f" status_message : {status_message} / {type(status_message)}")
-            # self.module.log(msg=f" status_message : {e.response.json()}")
 
             return status_code, status_message
 
         except ConnectionError as e:
             error_text = f"{type(e).__name__} {(str(e) if len(e.args) == 0 else str(e.args[0]))}"
             self.module.log(msg=f"ERROR   : {error_text}")
-            # self.module.log(msg="------------------------------------------------------------------")
             return 500, error_text
 
         except Exception as e:
             self.module.log(msg=f"ERROR   : {e}")
-            # self.module.log(msg=f" text    : {response.text} / {type(response.text)}")
-            # self.module.log(msg=f" json    : {response.json()} / {type(response.json())}")
-            # self.module.log(msg=f" headers : {response.headers}")
-            # self.module.log(msg=f" code    : {response.status_code}")
-            # self.module.log(msg="------------------------------------------------------------------")
 
             return response.status_code, response.text
 
@@ -259,9 +249,7 @@ class PHPComposerInstaller(object):
         """
         """
         rc, out, err = self.module.run_command(args, check_rc=True)
-        # self.module.log(msg="  rc : '{}'".format(rc))
-        # self.module.log(msg="  out: '{}' ({})".format(out, type(out)))
-        # self.module.log(msg="  err: '{}'".format(err))
+
         return rc, out, err
 
 
@@ -271,26 +259,36 @@ def main():
     argument_spec = dict(
         signature=dict(
             required=False,
-            default="https://composer.github.io/installer.sig"
+            default="https://composer.github.io/installer.sig",
+            type=str
         ),
         installer=dict(
             required=False,
-            default="https://getcomposer.org/installer"
+            default="https://getcomposer.org/installer",
+            type=str
         ),
         version=dict(
             required=False,
-            default=''
+            default='',
+            type=str
         ),
         version_branch=dict(
             required=False,
-            default='--2'
+            default='--2',
+            type=str
         ),
         self_update=dict(
             required=False,
-            default=False
+            default=False,
+            type=bool
         ),
         path=dict(
-            default='/usr/local/bin'
+            default='/usr/local/bin',
+            type=str
+        ),
+        force=dict(
+            default=False,
+            type=bool
         ),
     )
 
@@ -299,7 +297,7 @@ def main():
         supports_check_mode=False,
     )
 
-    helper = PHPComposerInstaller(module)
+    helper = ComposerInstaller(module)
     result = helper.run()
 
     module.log(msg=f" = result : '{result}'")
